UTs on 64bit OS, Part III
git-svn-id: https://svn.code.sf.net/p/quex/code/trunk@4614 db9deaaf-66c7-40b0-811a-ad7aae109365

+++ quex/code_base/buffer/lexatoms/converter/TEST-SUPPORT/basic_functionality.c
@@ -237,7 +237,8 @@ file_get_reference(const char* Codec)
         snprintf(&file_name[0], 1023, "%s.dat", file_stem);
     }
     else {
-        snprintf(&file_name[0], 1023, "%s-%i-%s.dat", file_stem, sizeof(QUEX_TYPE_LEXATOM)*8, 
+        snprintf(&file_name[0], 1023, "%s-%i-%s.dat", file_stem, 
+                 (int)sizeof(QUEX_TYPE_LEXATOM)*8, 
                  QUEXED(system_is_little_endian)() ? "le" : "be");
     }
     self.reference_file = &file_name[0];
